fix: 조회 입력 경계 검증 보완
- 브랜드와 상품 경로 ID가 0 이하이면 DB 조회 전에 400으로 거절
- 상품 등록의 brandId 누락과 양수 범위를 HTTP 입력 경계에서 검증
- page와 size의 offset 계산을 long으로 확장해 큰 페이지에서도 안전하게 처리
- 실제 HTTP 테스트로 잘못된 경로 ID와 기존 조회 흐름의 회귀를 검증

+++ apps/commerce-api/src/main/java/com/loopers/application/common/PageCriteria.java
@@ -10,7 +10,7 @@ public record PageCriteria(int page, int size) {
         }
     }
 
-    public int offset() {
-        return Math.multiplyExact(page, size);
+    public long offset() {
+        return Math.multiplyExact((long) page, size);
     }
 }
